デザインを調整
