--- web/css/main.css
@@ -1,48 +1,0 @@
-/* デフォルトのCSSをリセット */
-html body {
-    margin: 0;
-    padding: 0;
-}
-
-/* 要素を横並びに表示 */
-header {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 50px;
-}
-header h3 {
-    margin-right: auto;
-}
-
-body {
-    position: relative;
-    height: calc(100vh - 50px);
-}
-#rooplpp_program {
-    position: absolute;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    z-index: 0;
-}
-#result.deactive {
-    display: none;
-}
-#result.active {
-    position: absolute;
-    z-index: 1;
-    width: 100%;
-    height: 30vh;
-    bottom: -43px;
-}
-#result.active textarea {
-    width: 100%;
-    height: 90%;
-}
-#result.active button {
-    height: 9%;
-    margin: 0;
-    padding: 0;
-}
