+++ web/index.html
@@ -6,7 +6,6 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>ROOPL++ Playground</title>
 
-  <link rel="stylesheet" href="./css/main.css">
   <script src="./dist/main.bundle.js"></script>
 </head>
