+++ web/package.json
@@ -13,19 +13,24 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@material-ui/core": "^4.11.3",
+    "@material-ui/icons": "^4.11.2",
     "@types/ace": "0.0.44",
     "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
     "ace-builds": "^1.4.12",
     "axios": "^0.21.0",
+    "css-loader": "^5.2.0",
     "react": "^17.0.1",
     "react-ace": "^9.1.4",
     "react-dom": "^17.0.1",
+    "sass": "^1.32.8",
+    "sass-loader": "^11.0.1",
+    "style-loader": "^2.0.0",
     "ts-loader": "^8.0.7",
     "typescript": "^4.0.3",
     "webpack": "^5.2.0",
     "webpack-cli": "^4.1.0",
     "webpack-dev-server": "^3.11.0"
-  },
-  "dependencies": {}
+  }
 }
