+++ web/src/scss/foundation/_base.scss
@@ -1,0 +1,9 @@
+html,
+body {
+    height: 100%;
+    margin: 0;
+}
+
+#root {
+    height: 100%;
+}
