+++ web/src/scss/index.scss
@@ -1,0 +1,54 @@
+@charset "utf-8";
+
+// ==========================================================================
+// Foundation
+// ==========================================================================
+
+@use "foundation/_reset";
+@use "foundation/_base";
+
+// ==========================================================================
+// Layout
+// ==========================================================================
+
+
+// ==========================================================================
+// Object
+// ==========================================================================
+
+// -----------------------------------------------------------------
+// Component
+// -----------------------------------------------------------------
+
+@use 'object/component/_ace-editor';
+
+// -----------------------------------------------------------------
+// Project
+// -----------------------------------------------------------------
+
+
+// -----------------------------------------------------------------
+// Utility
+// -----------------------------------------------------------------
+
+@use 'object/utility/_margin';
+
+#result.deactive {
+    display: none;
+}
+#result.active {
+    position: absolute;
+    z-index: 1;
+    width: 100%;
+    height: 30vh;
+    bottom: -43px;
+}
+#result.active textarea {
+    width: 100%;
+    height: 90%;
+}
+#result.active button {
+    height: 9%;
+    margin: 0;
+    padding: 0;
+}
