+++ web/src/scss/object/component/_ace-editor.scss
@@ -1,0 +1,5 @@
+#ace-editor {
+    width: 100%;
+    height: -webkit-calc(100% - 64px);
+    height: calc(100% - 64px);
+}
