+++ web/src/scss/object/utility/_border-radius.scss
@@ -1,0 +1,18 @@
+/* ==========================================================================
+   Object
+   ========================================================================== */
+
+/* Utility
+   ----------------------------------------------------------------- */
+
+/**
+ * 角丸
+ */
+
+.border-radius--sm {
+    border-radius: 3px;
+}
+
+.border-radius--xl {
+    border-radius: 20px;
+}
