+++ web/src/scss/object/utility/_border.scss
@@ -1,0 +1,14 @@
+/* ==========================================================================
+   Object
+   ========================================================================== */
+
+/* Utility
+   ----------------------------------------------------------------- */
+
+/**
+ * ボーダー
+ */
+.border--frame {
+  border: 1px solid rgba(0, 0, 0, 0.2);
+  border-radius: 3px;
+}
