+++ web/src/scss/object/utility/_color.scss
@@ -1,0 +1,47 @@
+/* ==========================================================================
+   Object
+   ========================================================================== */
+
+/* Utility
+   ----------------------------------------------------------------- */
+
+/**
+ * 色
+ */
+.color-deep-black {
+    color: #333333;
+}
+.background-color-deep-black {
+    color: #333333;
+}
+
+.color-black {
+    color: #666666;
+}
+.background-color-black {
+    background-color: #666666;
+}
+
+.color-yellow {
+    color: #fffde3;
+}
+.background-color-yellow {
+    background-color: #fffde3;
+}
+.background-color-yellow--important {
+    background-color: #fffde3 !important;
+}
+
+.color-grey {
+    color: #f1f1f1;
+}
+.background-color-grey {
+    background-color: #f1f1f1;
+}
+
+.color-white {
+    color: #ffffff;
+}
+.background-color-white {
+    background-color: #ffffff;
+}
