+++ web/src/scss/object/utility/_padding.scss
@@ -1,0 +1,103 @@
+/* ==========================================================================
+   Object
+   ========================================================================== */
+
+/* Utility
+   ----------------------------------------------------------------- */
+
+/**
+ * padding
+ */
+// 上方向
+.padding-top-sm {
+  padding-top: 5px;
+}
+.padding-top-md {
+  padding-top: 10px;
+}
+.padding-top-lg {
+  padding-top: 15px;
+}
+.padding-top-xl {
+  padding-top: 20px;
+}
+.padding-top-xxl {
+  padding-top: 25px;
+}
+.padding-top-xxxl {
+  padding-top: 30px;
+}
+
+// 下方向
+.padding-bottom-sm {
+  padding-bottom: 5px;
+}
+.padding-bottom-md {
+  padding-bottom: 10px;
+}
+.padding-bottom-lg {
+  padding-bottom: 15px;
+}
+.padding-bottom-xl {
+  padding-bottom: 20px;
+}
+.padding-bottom-xxl {
+  padding-bottom: 25px;
+}
+.padding-bottom-xxxl {
+  padding-bottom: 30px;
+}
+
+// 右
+.padding-right-sm {
+  padding-right: 5px;
+}
+.padding-right-md {
+  padding-right: 10px;
+}
+.padding-right-lg {
+  padding-right: 15px;
+}
+.padding-right-xl {
+  padding-right: 20px;
+}
+.padding-right-xxl {
+  padding-right: 25px;
+}
+.padding-right-xxxl {
+  padding-right: 30px;
+}
+
+// 左
+.padding-left-sm {
+  padding-left: 5px;
+}
+.padding-left-md {
+  padding-left: 10px;
+}
+.padding-left-lg {
+  padding-left: 15px;
+}
+.padding-left-xl {
+  padding-left: 20px;
+}
+.padding-left-xxl {
+  padding-left: 25px;
+}
+.padding-left-xxxl {
+  padding-left: 30px;
+}
+
+// 周り
+.padding-box-sm {
+  padding: 5px;
+}
+.padding-box-md {
+  padding: 10px;
+}
+.padding-box-lg {
+  padding: 15px;
+}
+.padding-box-xl {
+  padding: 20px;
+}
