+++ web/src/scss/object/utility/_position.scss
@@ -1,0 +1,14 @@
+/* ==========================================================================
+   Object
+   ========================================================================== */
+
+/* Utility
+   ----------------------------------------------------------------- */
+
+/**
+ * ポジション
+ */
+
+ .relative {
+     position: relative;
+ }
